use separate query for org billing mode
