+++ components/dashboard/src/teams/OrgSettingsPage.tsx
@@ -4,14 +4,16 @@
  * See License.AGPL.txt in the project root for license information.
  */
 
+import { Team } from "@gitpod/gitpod-protocol";
+import { BillingMode } from "@gitpod/gitpod-protocol/lib/billing-mode";
 import { useMemo } from "react";
 import { Redirect } from "react-router";
 import Header from "../components/Header";
 import { SpinnerLoader } from "../components/Loader";
 import { PageWithSubMenu } from "../components/PageWithSubMenu";
 import { useFeatureFlags } from "../contexts/FeatureFlagContext";
+import { useOrgBillingMode } from "../data/billing-mode/org-billing-mode-query";
 import { useCurrentOrg } from "../data/organizations/orgs-query";
-import { getTeamSettingsMenu } from "./TeamSettings";
 
 export interface OrgSettingsPageProps {
     children: React.ReactNode;
@@ -19,24 +21,25 @@ export interface OrgSettingsPageProps {
 
 export function OrgSettingsPage({ children }: OrgSettingsPageProps) {
     const org = useCurrentOrg();
+    const orgBillingMode = useOrgBillingMode();
     const { oidcServiceEnabled, orgGitAuthProviders } = useFeatureFlags();
 
     const menu = useMemo(
         () =>
             getTeamSettingsMenu({
                 team: org.data,
-                billingMode: org.data?.billingMode,
+                billingMode: orgBillingMode.data,
                 ssoEnabled: oidcServiceEnabled,
                 orgGitAuthProviders,
             }),
-        [oidcServiceEnabled, orgGitAuthProviders, org.data],
+        [org.data, orgBillingMode.data, oidcServiceEnabled, orgGitAuthProviders],
     );
 
     const title = "Organization Settings";
     const subtitle = "Manage your organization's settings.";
 
     // Render as much of the page as we can in a loading state to avoid content shift
-    if (org.isLoading) {
+    if (org.isLoading || orgBillingMode.isLoading) {
         return (
             <div className="w-full">
                 <Header title={title} subtitle={subtitle} />
@@ -57,4 +60,39 @@ export function OrgSettingsPage({ children }: OrgSettingsPageProps) {
             {children}
         </PageWithSubMenu>
     );
+}
+
+export function getTeamSettingsMenu(params: {
+    team?: Team;
+    billingMode?: BillingMode;
+    ssoEnabled?: boolean;
+    orgGitAuthProviders: boolean;
+}) {
+    const { billingMode, ssoEnabled, orgGitAuthProviders } = params;
+    const result = [
+        {
+            title: "General",
+            link: [`/settings`],
+        },
+    ];
+    if (ssoEnabled) {
+        result.push({
+            title: "SSO",
+            link: [`/sso`],
+        });
+    }
+    if (orgGitAuthProviders) {
+        result.push({
+            title: "Git Auth",
+            link: [`/settings/git`],
+        });
+    }
+    if (billingMode?.mode !== "none") {
+        // The Billing page contains both chargebee and usage-based components, so: always show them!
+        result.push({
+            title: "Billing",
+            link: ["/billing"],
+        });
+    }
+    return result;
 }
